Rough sample with player list and placed list
List containing players who have been voted out which can't be
rearranged and a list of remaining players who are still playing and can
be rearranged. Players from the playing list can be moved to the voted
out list. Player list is styles with tribe color, name and profile
picture.