Use Component Did Update to wire up sortable lists
Originally Component Did Mount was used to wire up the sortable lists.
But now since a loading page is displayed these lists don't exist to
start with and so Component Did Mount cannot wire up the lists to be
sortable. So now the lists are made sortable when the component updates.
This should be a cheap operation, but it would be ideal to only do it
once.